splitting the go project code
